Various improvements - Better field naming in DetailedCommandException - Included the Error that was actually thrown in Event.parseSnapshot - Null-check in Emitter - Renamed VM.platformOptions get/setter to VM.platform - Fixed long to int cast in VDI

+++ src/main/java/net/wgr/xenmaster/api/VDI.java
@@ -62,7 +62,7 @@ public class VDI extends NamedEntity {
     }
 
     public String create(double sizeInMb, Type type, SR repo, boolean shareable, boolean readOnly) throws BadAPICallException {
-        this.virtualSize = (int) (sizeInMb * (megabyte));
+        this.virtualSize = (long) (sizeInMb * (megabyte));
         this.type = type;
         this.sr = repo.getReference();
         this.shareable = shareable;
